lateNigth Change
So I added the three layers so that we can have things more organized. Ken, I disabled the idle animation for now cuz it was making the whole screen move, but just change the animation sizing(I THINK) and it should prevent it from doing that. Cuz on itch it wasnt a problem so its weird that its happening now. I added a button that unlocks the light blue door. the goal of this first layouy should be to teach the player the basic mechanics that we use. I wanna add an enemy in this layer, he should be slowly moving by the exit or just stuck in place. (Shooting at the player if we decide to implement them shooting the boogies since they're sick). I wanna show the player that they can shoot the arm at it or simply avoid them. Later on Imma put up a list of what everyone should be doing for this week so that everyone knows what they're doing and the game gets polished :)